requirements.txt
streamlit==1.24.0
google.generativeai
python-dotenv